organize database related files
