+++ app/persistence/utils/polygon_utils.py
